efficient model loading

+++ hdm2/utils/model_utils.py
@@ -289,8 +289,8 @@ def _detect_hallucinations(prompt, context, response,
         op = token_model(**model_inputs)
     
     # Use original logic for token scores and sequence probabilities
-    token_scores = torch.softmax(op['token_logits'][0], axis=1)[prompt_len:, 1].cpu().numpy()
-    seq_probs = torch.softmax(op['seq_logits'], dim=-1)[0].cpu().numpy()
+    token_scores = torch.softmax(op['token_logits'][0], axis=1)[prompt_len:, 1].float().cpu().numpy()
+    seq_probs = torch.softmax(op['seq_logits'], dim=-1)[0].float().cpu().numpy()
     
     # Extract sentences
     sentences_data = prepare_sentences(response)
